requirements.txt
pandas
numpy
timm
torch
torchvision
megadetector

pillow
fastapi
uvicorn
python-multipart
